SMS-249

+++ homework-service/src/main/java/com/sms/homeworkservice/homeworks/control/repository/HomeworksFilesRepository.java
@@ -12,9 +12,8 @@ public interface HomeworksFilesRepository extends CrudRepository<FileJPA, Long> 
     @NonNull
     List<FileJPA> findAll();
 
-    List<FileJPA> findAllById(Integer id);
+    List<FileJPA> findAllById(Long id);
 
     List<FileJPA> findAllByHomeworkid(Integer homeworkID);
-
 
 }
